+++ app/scripts/models/app-model.js
@@ -88,6 +88,9 @@ const AppModel = Backbone.Model.extend({
     applyUserConfig(config) {
         this.settings.set(config.settings);
         if (config.files) {
+            if (config.showOnlyFilesFromConfig) {
+                this.fileInfos.reset();
+            }
             config.files
                 .filter(file => file && file.storage && file.name && file.path &&
                     !this.fileInfos.getMatch(file.storage, file.name, file.path))