+++ app/scripts/app.js
@@ -72,7 +72,7 @@ ready(() => {
             if (configParam) {
                 appModel.loadConfig(configParam, err => {
                     SettingsManager.setBySettings(appModel.settings);
-                    if (err) {
+                    if (err && !appModel.settings.get('cacheConfigSettings')) {
                         showSettingsLoadError();
                         reject(err);
                     } else {