+++ release-notes.md
@@ -4,6 +4,7 @@ Release notes
 `+` file path hint in recent files list  
 `+` cacheConfigSettings config option  
 `+` keyboard-accessible autocomplete  
+`+` entry auto-type context menu  
 
 ##### v1.4.0 (2017-02-04)
 KDBX4 format support and minor improvements