+++ release-notes.md
@@ -5,6 +5,7 @@ Release notes
 `+` cacheConfigSettings config option  
 `+` keyboard-accessible autocomplete  
 `+` entry auto-type context menu  
+`+` `kw.require` now provides access to modules  
 
 ##### v1.4.0 (2017-02-04)
 KDBX4 format support and minor improvements